+++ lib/src/interfaces.dart
@@ -1,4 +1,4 @@
-import 'job_info.dart';
+import 'async_node.dart';
 import 'typedef.dart';
 
 abstract class AsyncQueueInterface {
@@ -8,7 +8,8 @@ abstract class AsyncQueueInterface {
   void retry();
   void addJob(AsyncJob job, {String? label, int retryTime});
   void addJobThrow(AsyncJob job);
+  void addNode(AsyncNode node);
   Future<void> start();
-  List<JobInfo> list();
-  JobInfo getJobInfo(String label);
+  List<AsyncNode> list();
+  AsyncNode getJob(String label);
 }